Move text rendering to separate file
